+++ src/Symfony/Component/Messenger/Tests/Command/FailedMessagesShowCommandTest.php
@@ -15,6 +15,7 @@ use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Tester\CommandTester;
 use Symfony\Component\Messenger\Command\FailedMessagesShowCommand;
 use Symfony\Component\Messenger\Envelope;
+use Symfony\Component\Messenger\Stamp\ErrorDetailsStamp;
 use Symfony\Component\Messenger\Stamp\RedeliveryStamp;
 use Symfony\Component\Messenger\Stamp\SentToFailureTransportStamp;
 use Symfony\Component\Messenger\Stamp\TransportMessageIdStamp;
@@ -29,11 +30,13 @@ class FailedMessagesShowCommandTest extends TestCase
     public function testBasicRun()
     {
         $sentToFailureStamp = new SentToFailureTransportStamp('async');
-        $redeliveryStamp = new RedeliveryStamp(0, 'Things are bad!');
+        $redeliveryStamp = new RedeliveryStamp(0);
+        $errorStamp = new ErrorDetailsStamp(new \Exception('Things are bad!', 123));
         $envelope = new Envelope(new \stdClass(), [
             new TransportMessageIdStamp(15),
             $sentToFailureStamp,
             $redeliveryStamp,
+            $errorStamp,
         ]);
         $receiver = $this->createMock(ListableReceiverInterface::class);
         $receiver->expects($this->once())->method('find')->with(15)->willReturn($envelope);
@@ -52,8 +55,9 @@ class FailedMessagesShowCommandTest extends TestCase
   Message Id    15                   
   Failed at     %s  
   Error         Things are bad!      
-  Error Class   (unknown)            
-  Transport     async
+  Error Code    123                  
+  Error Class   Exception            
+  Transport     async                
 EOF
             ,
             $redeliveryStamp->getRedeliveredAt()->format('Y-m-d H:i:s')),
@@ -63,36 +67,71 @@ EOF
     public function testMultipleRedeliveryFails()
     {
         $sentToFailureStamp = new SentToFailureTransportStamp('async');
-        $redeliveryStamp1 = new RedeliveryStamp(0, 'Things are bad!');
+        $redeliveryStamp1 = new RedeliveryStamp(0);
+        $errorStamp = new ErrorDetailsStamp(new \Exception('Things are bad!', 123));
         $redeliveryStamp2 = new RedeliveryStamp(0);
         $envelope = new Envelope(new \stdClass(), [
             new TransportMessageIdStamp(15),
             $sentToFailureStamp,
             $redeliveryStamp1,
+            $errorStamp,
             $redeliveryStamp2,
         ]);
         $receiver = $this->createMock(ListableReceiverInterface::class);
         $receiver->expects($this->once())->method('find')->with(15)->willReturn($envelope);
-
         $command = new FailedMessagesShowCommand(
             'failure_receiver',
             $receiver
         );
-
         $tester = new CommandTester($command);
         $tester->execute(['id' => 15]);
-
         $this->assertStringContainsString(sprintf(<<<EOF
  ------------- --------------------- 
   Class         stdClass             
   Message Id    15                   
   Failed at     %s  
   Error         Things are bad!      
-  Error Class   (unknown)            
-  Transport     async
+  Error Code    123                  
+  Error Class   Exception            
+  Transport     async                
 EOF
             ,
             $redeliveryStamp2->getRedeliveredAt()->format('Y-m-d H:i:s')),
+            $tester->getDisplay(true));
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testLegacyFallback()
+    {
+        $sentToFailureStamp = new SentToFailureTransportStamp('async');
+        $redeliveryStamp = new RedeliveryStamp(0, 'Things are bad!');
+        $envelope = new Envelope(new \stdClass(), [
+            new TransportMessageIdStamp(15),
+            $sentToFailureStamp,
+            $redeliveryStamp,
+        ]);
+        $receiver = $this->createMock(ListableReceiverInterface::class);
+        $receiver->expects($this->once())->method('find')->with(15)->willReturn($envelope);
+        $command = new FailedMessagesShowCommand(
+            'failure_receiver',
+            $receiver
+        );
+        $tester = new CommandTester($command);
+        $tester->execute(['id' => 15]);
+        $this->assertStringContainsString(sprintf(<<<EOF
+ ------------- --------------------- 
+  Class         stdClass             
+  Message Id    15                   
+  Failed at     %s  
+  Error         Things are bad!      
+  Error Code                         
+  Error Class   (unknown)            
+  Transport     async                
+EOF
+            ,
+            $redeliveryStamp->getRedeliveredAt()->format('Y-m-d H:i:s')),
             $tester->getDisplay(true));
     }
 
@@ -113,11 +152,13 @@ EOF
     public function testListMessages()
     {
         $sentToFailureStamp = new SentToFailureTransportStamp('async');
-        $redeliveryStamp = new RedeliveryStamp(0, 'Things are bad!');
+        $redeliveryStamp = new RedeliveryStamp(0);
+        $errorStamp = new ErrorDetailsStamp(new \RuntimeException('Things are bad!'));
         $envelope = new Envelope(new \stdClass(), [
             new TransportMessageIdStamp(15),
             $sentToFailureStamp,
             $redeliveryStamp,
+            $errorStamp,
         ]);
         $receiver = $this->createMock(ListableReceiverInterface::class);
         $receiver->expects($this->once())->method('all')->with()->willReturn([$envelope]);
@@ -130,7 +171,7 @@ EOF
         $tester = new CommandTester($command);
         $tester->execute([]);
         $this->assertStringContainsString(sprintf(<<<EOF
-15   stdClass   %s   Things are bad!  
+15   stdClass   %s   Things are bad!
 EOF
             ,
             $redeliveryStamp->getRedeliveredAt()->format('Y-m-d H:i:s')),
@@ -158,7 +199,8 @@ EOF
         $envelope = new Envelope(new \stdClass(), [
             new TransportMessageIdStamp(15),
             $sentToFailureStamp,
-            new RedeliveryStamp(0, 'Things are bad!'),
+            new RedeliveryStamp(0),
+            new ErrorDetailsStamp(new \RuntimeException('Things are bad!')),
         ]);
         $receiver = $this->createMock(ListableReceiverInterface::class);
         $receiver->expects($this->once())->method('all')->with()->willReturn([$envelope]);